Get __shared__ for onReady
