+++ library/src/main/java/com/paypal/messages/ModalFragment.kt
@@ -29,6 +29,7 @@ import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.google.android.material.bottomsheet.BottomSheetDialog
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.google.gson.Gson
+import com.google.gson.GsonBuilder
 import com.google.gson.JsonParser
 import com.paypal.messages.analytics.AnalyticsComponent
 import com.paypal.messages.analytics.AnalyticsEvent
@@ -53,6 +54,7 @@ internal class ModalFragment(
 ) : BottomSheetDialogFragment() {
 	private val TAG = "PayPalMessageModal"
 	private val offsetTop = 50.dp
+	private val gson = GsonBuilder().setPrettyPrinting().create()
 
 	private var modalUrl: String? = null
 
@@ -371,7 +373,7 @@ internal class ModalFragment(
 		val nameAndArgs = JsonParser.parseString(params).asJsonObject
 		val name = nameAndArgs.get("name").asString
 		val args = nameAndArgs.get("args").asJsonArray[0].asJsonObject
-		LogCat.debug(TAG, "CallbackHandler:\n  name = $name\n  args = $args")
+		LogCat.debug(TAG, "CallbackHandler:\n  name = $name\n  args = ${gson.toJson(args)}")
 
 		// If __shared__ does not exist, use an empty object
 		val sharedJson = args.get("__shared__") ?: JsonParser.parseString("{}")
@@ -403,6 +405,17 @@ internal class ModalFragment(
 					AnalyticsEvent(
 						eventType = EventType.MODAL_CLICKED,
 						data = "$calculatorAmount",
+					),
+					shared,
+				)
+			}
+
+			"onReady" -> {
+				logEvent(
+					AnalyticsEvent(
+						eventType = EventType.MESSAGE_RENDERED,
+						renderDuration = args.get("render_duration")?.asString ?: "",
+						requestDuration = args.get("request_duration")?.asString ?: "",
 					),
 					shared,
 				)
